requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
charset-normalizer==3.3.1
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==6.0.0
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
etelemetry==0.3.1
filelock==3.13.0
fitz==0.0.1.dev2
future==0.18.3
httplib2==0.22.0
idna==3.4
importlib-resources==6.1.0
isodate==0.6.1
looseversion==1.3.0
lxml==4.9.3
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
numpy==1.24.4
packaging==23.2
pandas==2.0.3
pathlib==1.0.1
Pillow==10.1.0
prov==2.0.0
pydot==1.4.2
PyMuPDF==1.23.5
PyMuPDFb==1.23.5
pyparsing==3.1.1
PyPDF2~=2.0
python-dateutil==2.8.2
pytz==2023.3.post1
pyxnat==1.6
rdflib==7.0.0
reportlab==4.0.6
requests==2.31.0
scipy==1.10.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
traits==6.3.2
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
zipp==3.17.0